+++ CHANGELOG.md
@@ -7,6 +7,7 @@ by S. Gariazzo (gariazzo@ific.uv.es), P.F. de Salas (pablo.fernandez@fysik.su.se
 
 ## v1.1.0 ()
 * neutrino-neutrino collision terms take into account the presence of sterile neutrinos
+* muons are now disabled by default, enable their contributions compiling with DO_MUONS=1
 * first (very incomplete) python wrapper for the fortran code
 * possibility to store intermediate quantities for later checks (the overall normalization, Y, Ydot, Heff, the commutator, collision terms as functions of x)
 * Makefile now defaults to gfortran